ProyectoLenguajes/requirements.txt
Django==5.2.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
requests==2.31.0

